moved reusables to another file to be used for teacher dashboard
